Addresses #50: Support unsigned 64-bit integers for sequence numbers
There's no internal arithmetic or other operations that use the values
of SeqNum fields.  And Python's integers will scale transparently from
fixed length to arbitrary length as required.

Added test cases for both encoder and decoder, checking that it can
handle values up to 2**64, which (per the bug report) is a billion
messages a second for 584 years, so ... all good.